updated to release v1.11.4

See merge request palisade/palisade-release!18

+++ src/core/include/lattice/trapdoor.h
@@ -279,7 +279,7 @@ class RLWETrapdoorUtility {
       }
     } else {
       // Peikert's inversion sampling method
-      std::shared_ptr<int32_t> dggVector =
+      std::shared_ptr<int64_t> dggVector =
           dggLargeSigma.GenerateIntVector(n * k);
       for (size_t i = 0; i < n * k; i++) {
         p2ZVector(i, 0) = (dggVector.get())[i];
@@ -387,7 +387,7 @@ class RLWETrapdoorUtility {
       }
     } else {
       // Peikert's inversion sampling method
-      std::shared_ptr<int32_t> dggVector =
+      std::shared_ptr<int64_t> dggVector =
           dggLargeSigma.GenerateIntVector(n * k * d);
 
       for (size_t i = 0; i < n * k; i++) {
